Fix: Deprecated LevelManager.js Chore: Moved platformer utilities into platformer directory Chore: Updated platformer utility paths
