+++ src/levels/platformer/entities/Platform.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import PhysicsObject from '../../../utils/PhysicsObject';
+import PhysicsObject from '../../../levels/platformer/utils/PhysicsObject';
 /* A temp class to create basic platforms using box geometry */
 class Platform {
     constructor(scene, x, y, z, width, height, depth, physicsEngine, id) {
