+++ src/levels/platformer/utils/AssetLoader.js
